refactor: some adjusts :hammer:
