+++ components/transaction/internal/gen/inject.go
@@ -6,6 +6,7 @@ package gen
 import (
 	"fmt"
 	"github.com/LerianStudio/midaz/components/transaction/internal/adapters/database/postgres"
+	"github.com/LerianStudio/midaz/components/transaction/internal/ports"
 	"sync"
 
 	"github.com/LerianStudio/midaz/common"
@@ -59,7 +60,7 @@ var (
 		httpHandler.NewRouter,
 		service.NewServer,
 		postgres.NewTransactionPostgreSQLRepository,
-		wire.Struct(new(httpHandler.TransactionHandler), "*"),
+		wire.Struct(new(ports.TransactionHandler), "*"),
 		wire.Struct(new(command.UseCase), "*"),
 		wire.Struct(new(query.UseCase), "*"),
 		wire.Bind(new(t.Repository), new(*postgres.TransactionPostgreSQLRepository)),
