+++ components/transaction/internal/ports/transaction.go
@@ -1,4 +1,4 @@
-package http
+package ports
 
 import (
 	"bytes"
